Change Pixel Ratio and adjust heights for models
Experiment with setting the game to PixelRatio 1.0.

This code will automatically resize model-based actors, and their associated invisible bridges where applicable, to fit the new size of the world.

When PixelRatio is 1.0, the ModelHeightMul constant should be set to 1.2.
If it is decided to revert the world to PixelRatio 1.2, setting ModelHeightMul to 1.0 will ensure the actors are the right size again.