Tweak Pipe5 and its derivatives
Pixelratio made it clear that these pipes' heights need to change with the world, so scaling them down is necessary, and to keep them from getting oblong, I'll scale them on all dimensions. This makes them shrink away from the walls, however, so I'm having the actors scoot toward the wall to compensate.