Convert PseudoMusic into actual Music
PseudoMusic was done, far as I can tell, to allow the music to be paused (so it doesn't run longer than expected) and to prevent looping. There are legitimate ways to do this with MAPINFO and ZScript, so let's go with that. Now the engine's music system can deal with them instead, and deep water won't distort the songs anymore.